implement basic tests Calculator class, use BigInteger as a return type (to operate on huge rectangles)

+++ src/main/java/com/jarqprog/calculator/OverlapCalc.java
@@ -1,6 +1,8 @@
 package com.jarqprog.calculator;
 
+import java.math.BigInteger;
+
 public interface OverlapCalc {
 
-    long calculateOverlapArea(int[] firstRectangleCoordinates, int[] secondRectangleCoordinates);
+    BigInteger calculateOverlapArea(int[] firstRectangleCoordinates, int[] secondRectangleCoordinates);
 }
